Move expression UID constant to types
