+++ public/app/features/alerting/unified/utils/query.test.ts
@@ -1,5 +1,5 @@
 import { DataSourceJsonData, PluginMeta } from '@grafana/data';
-import { ExpressionDatasourceUID } from 'app/features/expressions/ExpressionDatasource';
+import { ExpressionDatasourceUID } from 'app/features/expressions/types';
 import { CombinedRule } from 'app/types/unified-alerting';
 import { GrafanaAlertStateDecision } from 'app/types/unified-alerting-dto';
