+++ public/app/features/expressions/types.ts
@@ -2,6 +2,11 @@ import { DataQuery, ReducerID, SelectableValue } from '@grafana/data';
 
 import { EvalFunction } from '../alerting/state/alertDef';
 
+/**
+ * MATCHES a constant in DataSourceWithBackend
+ */
+export const ExpressionDatasourceUID = '__expr__';
+
 export enum ExpressionQueryType {
   math = 'math',
   reduce = 'reduce',
